Move key methods into module
